Improved: Updated to Visual Studio 2019 and Qt 5.15.2. Running Equalizer APO requires Universal CRT now, which means that Windows versions before 10 may need additional Windows updates or the Visual C++ Redistributable for Visual Studio 2015 - 2019 needs to be installed (installer provides assistance).
git-svn-id: svn://svn.code.sf.net/p/equalizerapo/code/trunk@83 177c5db2-39f6-4273-851a-b0754c4d5937

+++ parser/RegistryFunctions.cpp
@@ -85,7 +85,7 @@ void ReadRegDWORDFunction::Eval(ptr_val_type& ret, const ptr_val_type* arg, int 
 	{
 		unsigned long value = RegistryHelper::readDWORDValue(key, valuename);
 
-		*ret = (long)value;
+		*ret = (int)value;
 
 		engine->watchRegistryKey(key);
 	}
